Add a test for the case where FixityChecker has a locate() error
Testing the FixityChecker with coverage enabled showed that we didn't
have any tests for the branch

    locate(expectedFileFixity.uri) match {
      ...
      case Left(locateError) =>
    }

so check we do actually return the appropriate error in this branch.
